Enum minijail::Error

pub enum Error {
    BindMount {
        errno: i32,
        src: PathBuf,
        dst: PathBuf,
    },
    Mount {
        errno: i32,
        src: PathBuf,
        dest: PathBuf,
        fstype: String,
        flags: usize,
        data: String,
    },
    CheckingMultiThreaded(Error),
    CreatingMinijail,
    ForkingMinijail(i32),
    ForkingWhileMultiThreaded,
    SeccompPath(PathBuf),
    StrToCString(String),
    PathToCString(PathBuf),
    DupDevNull(i32),
    OpenDevNull(Error),
    ReadProgram(Error),
    SetAltSyscallTable {
        errno: i32,
        name: String,
    },
    SetRlimit {
        errno: i32,
        kind: c_int,
    },
    SettingChrootDirectory(i32, PathBuf),
    SettingPivotRootDirectory(i32, PathBuf),
    ReadFdDirEntry(Error),
    ReadFdDir(Error),
    ProcFd(String),
    PreservingFd(i32),
    ProgramTooLarge,
    WrongProgramAlignment,
    WrongProgramSize,
    NoCommand,
    NoAccess,
    SeccompViolation(i32),
    Killed(u8),
    ReturnCode(u8),
}

Variants

BindMount

Fields of BindMount

errno: i32src: PathBufdst: PathBuf

Mount

Fields of Mount

errno: i32src: PathBufdest: PathBuffstype: Stringflags: usizedata: String

CheckingMultiThreaded(Error)

Failure to count the number of threads in /proc/self/tasks.

CreatingMinijail

minjail_new failed, this is an allocation failure.

ForkingMinijail(i32)

minijail_fork failed with the given error code.

ForkingWhileMultiThreaded

Attempt to fork while already multithreaded.

SeccompPath(PathBuf)

The seccomp policy path doesn't exist.

StrToCString(String)

The string passed in didn't parse to a valid CString.

PathToCString(PathBuf)

The path passed in didn't parse to a valid CString.

DupDevNull(i32)

Failed to call dup2 to set stdin, stdout, or stderr to /dev/null.

OpenDevNull(Error)

Failed to set up /dev/null for FDs 0, 1, or 2.

ReadProgram(Error)

Failed to read policy bpf from file.

SetAltSyscallTable

Setting the specified alt-syscall table failed with errno. Is the table in the kernel?

Fields of SetAltSyscallTable

errno: i32name: String

SetRlimit

Setting the specified rlimit failed with errno.

Fields of SetRlimit

errno: i32kind: c_int

SettingChrootDirectory(i32, PathBuf)

chroot failed with the provided errno.

SettingPivotRootDirectory(i32, PathBuf)

pivot_root failed with the provided errno.

ReadFdDirEntry(Error)

There is an entry in /proc/self/fd that isn't a valid PID.

ReadFdDir(Error)

/proc/self/fd failed to open.

ProcFd(String)

An entry in /proc/self/fd is not an integer

PreservingFd(i32)

Minijail refused to preserve an FD in the inherit list of fork().

ProgramTooLarge

Program size is too large

WrongProgramAlignment

Alignment of file should be divisible by the alignment of sock_filter.

WrongProgramSize

File size should be non-zero and a multiple of sock_filter

NoCommand

The command was not found.

NoAccess

The command could not be run.

SeccompViolation(i32)

Process was killed by SIGSYS indicating a seccomp violation.

Killed(u8)

Process was killed by a signal other than SIGSYS.

ReturnCode(u8)

Process finished returning a non-zero code.