Struct minijail::Minijail

pub struct Minijail { /* fields omitted */ }

Configuration to jail a process based on wrapping libminijail.

Intentionally leave out everything related to minijail_run. Forking is hard to reason about w.r.t. memory and resource safety. It is better to avoid forking from rust code. Leave forking to the library user, who can make an informed decision about when to fork to minimize risk.

Examples

Load seccomp policy - like "minijail0 -n -S myfilter.policy"

      let mut j = Minijail::new().map_err(|_| ())?;
      j.no_new_privs();
      j.parse_seccomp_filters("my_filter.policy").map_err(|_| ())?;
      j.use_seccomp_filter();
      unsafe { // `fork` will close all the programs FDs.
          j.fork(None).map_err(|_| ())?;
      }

Keep stdin, stdout, and stderr open after jailing.

      let j = Minijail::new().map_err(|_| ())?;
      let preserve_fds: Vec<RawFd> = vec![0, 1, 2];
      unsafe { // `fork` will close all the programs FDs.
          j.fork(Some(&preserve_fds)).map_err(|_| ())?;
      }

Errors

The fork function might not return an error if it fails after forking. A partial jail is not recoverable and will instead result in killing the process.

Implementations

impl Minijail

pub fn new() -> Result<Minijail>

Creates a new jail configuration.

pub fn try_clone(&self) -> Result<Minijail>

Clones self to a new Minijail. Useful because fork can only be called once on a Minijail.

pub fn change_uid(&mut self, uid: uid_t)

pub fn change_gid(&mut self, gid: gid_t)

pub fn change_user(&mut self, user: &str) -> Result<()>

pub fn change_group(&mut self, group: &str) -> Result<()>

pub fn set_supplementary_gids(&mut self, ids: &[gid_t])

pub fn keep_supplementary_gids(&mut self)

pub fn set_rlimit(
&mut self,
kind: c_int,
cur: rlim_t,
max: rlim_t
) -> Result<()>

pub fn use_seccomp(&mut self)

pub fn no_new_privs(&mut self)

pub fn use_seccomp_filter(&mut self)

pub fn set_seccomp_filter_tsync(&mut self)

pub fn parse_seccomp_program<P: AsRef<Path>>(&mut self, path: P) -> Result<()>

pub fn parse_seccomp_filters<P: AsRef<Path>>(&mut self, path: P) -> Result<()>

pub fn log_seccomp_filter_failures(&mut self)

pub fn use_caps(&mut self, capmask: u64)

pub fn capbset_drop(&mut self, capmask: u64)

pub fn set_ambient_caps(&mut self)

pub fn reset_signal_mask(&mut self)

pub fn run_as_init(&mut self)

pub fn namespace_pids(&mut self)

pub fn namespace_user(&mut self)

pub fn namespace_user_disable_setgroups(&mut self)

pub fn namespace_vfs(&mut self)

pub fn new_session_keyring(&mut self)

pub fn skip_remount_private(&mut self)

pub fn namespace_ipc(&mut self)

pub fn namespace_net(&mut self)

pub fn namespace_cgroups(&mut self)

pub fn remount_proc_readonly(&mut self)

pub fn set_remount_mode(&mut self, mode: c_ulong)

pub fn uidmap(&mut self, uid_map: &str) -> Result<()>

pub fn gidmap(&mut self, gid_map: &str) -> Result<()>

pub fn inherit_usergroups(&mut self)

pub fn use_alt_syscall(&mut self, table_name: &str) -> Result<()>

pub fn enter_chroot<P: AsRef<Path>>(&mut self, dir: P) -> Result<()>

pub fn enter_pivot_root<P: AsRef<Path>>(&mut self, dir: P) -> Result<()>

pub fn mount<P1: AsRef<Path>, P2: AsRef<Path>>(
&mut self,
src: P1,
dest: P2,
fstype: &str,
flags: usize
) -> Result<()>

pub fn mount_with_data<P1: AsRef<Path>, P2: AsRef<Path>>(
&mut self,
src: P1,
dest: P2,
fstype: &str,
flags: usize,
data: &str
) -> Result<()>

pub fn mount_dev(&mut self)

pub fn mount_tmp(&mut self)

pub fn mount_tmp_size(&mut self, size: usize)

pub fn mount_bind<P1: AsRef<Path>, P2: AsRef<Path>>(
&mut self,
src: P1,
dest: P2,
writable: bool
) -> Result<()>

pub fn run<P: AsRef<Path>, S: AsRef<str>>(
&self,
cmd: P,
inheritable_fds: &[RawFd],
args: &[S]
) -> Result<pid_t>

Forks and execs a child and puts it in the previously configured minijail. FDs 0, 1, and 2 are overwritten with /dev/null FDs unless they are included in the inheritable_fds list. This function may abort in the child on error because a partially entered jail isn't recoverable.

pub fn run_remap<P: AsRef<Path>, S: AsRef<str>>(
&self,
cmd: P,
inheritable_fds: &[(RawFd, RawFd)],
args: &[S]
) -> Result<pid_t>

Behaves the same as run() except inheritable_fds is a list of fd mappings rather than just a list of fds to preserve.

pub fn run_fd<F: AsRawFd, S: AsRef<str>>(
&self,
cmd: &F,
inheritable_fds: &[RawFd],
args: &[S]
) -> Result<pid_t>

Behaves the same as run() except cmd is a file descriptor to the executable.

pub fn run_fd_remap<F: AsRawFd, S: AsRef<str>>(
&self,
cmd: &F,
inheritable_fds: &[(RawFd, RawFd)],
args: &[S]
) -> Result<pid_t>

Behaves the same as run() except cmd is a file descriptor to the executable, and inheritable_fds is a list of fd mappings rather than just a list of fds to preserve.

pub unsafe fn fork(&self, inheritable_fds: Option<&[RawFd]>) -> Result<pid_t>

Forks a child and puts it in the previously configured minijail.

Safety

fork is unsafe because it closes all open FD for this process. That could cause a lot of trouble if not handled carefully. FDs 0, 1, and 2 are overwritten with /dev/null FDs unless they are included in the inheritable_fds list.

Also, any Rust objects that own fds may try to close them after the fork. If they belong to a fd number that was mapped to, the mapped fd will be closed instead.

This Function may abort in the child on error because a partially entered jail isn't recoverable.

pub unsafe fn fork_remap(
&self,
inheritable_fds: &[(RawFd, RawFd)]
) -> Result<pid_t>

Behaves the same as fork() except inheritable_fds is a list of fd mappings rather than just a list of fds to preserve.

Safety

See fork.

pub fn wait(&self) -> Result<()>

pub fn kill(&self) -> Result<()>

Send a SIGTERM to the child process and wait for its return code.