Enum minijail::Error

pub enum Error {

    BindMount {
        errno: i32,
        src: PathBuf,
        dst: PathBuf,
    },
    Mount {
        errno: i32,
        src: PathBuf,
        dest: PathBuf,
        fstype: String,
        flags: usize,
        data: String,
    },
    CheckingMultiThreaded(Error),
    CreatingMinijail,
    ForkingMinijail(i32),
    ForkingWhileMultiThreaded,
    SeccompPath(PathBuf),
    StrToCString(String),
    PathToCString(PathBuf),
    DupDevNull(i32),
    OpenDevNull(Error),
    ReadProgram(Error),
    SetAltSyscallTable {
        errno: i32,
        name: String,
    },
    SetRlimit {
        errno: i32,
        kind: c_int,
    },
    SettingChrootDirectory(i32, PathBuf),
    SettingPivotRootDirectory(i32, PathBuf),
    ReadFdDirEntry(Error),
    ReadFdDir(Error),
    ProcFd(String),
    PreservingFd(i32),
    ProgramTooLarge,
    WrongProgramAlignment,
    WrongProgramSize,
    NoCommand,
    NoAccess,
    SeccompViolation(i32),
    Killed(u8),
    ReturnCode(u8),
    Wait(i32),

}

Variants

BindMount

Fields of BindMount

errno: i32src: PathBufdst: PathBuf

Mount

Fields of Mount

errno: i32src: PathBufdest: PathBuffstype: Stringflags: usizedata: String

CheckingMultiThreaded(Error)

Failure to count the number of threads in /proc/self/tasks.

Tuple Fields of CheckingMultiThreaded

0: Error

CreatingMinijail

minjail_new failed, this is an allocation failure.

ForkingMinijail(i32)

minijail_fork failed with the given error code.

Tuple Fields of ForkingMinijail

0: i32

ForkingWhileMultiThreaded

Attempt to fork while already multithreaded.

SeccompPath(PathBuf)

The seccomp policy path doesn’t exist.

Tuple Fields of SeccompPath

0: PathBuf

StrToCString(String)

The string passed in didn’t parse to a valid CString.

Tuple Fields of StrToCString

0: String

PathToCString(PathBuf)

The path passed in didn’t parse to a valid CString.

Tuple Fields of PathToCString

0: PathBuf

DupDevNull(i32)

Failed to call dup2 to set stdin, stdout, or stderr to /dev/null.

Tuple Fields of DupDevNull

0: i32

OpenDevNull(Error)

Failed to set up /dev/null for FDs 0, 1, or 2.

Tuple Fields of OpenDevNull

0: Error

ReadProgram(Error)

Failed to read policy bpf from file.

Tuple Fields of ReadProgram

0: Error

SetAltSyscallTable

Setting the specified alt-syscall table failed with errno. Is the table in the kernel?

Fields of SetAltSyscallTable

errno: i32name: String

SetRlimit

Setting the specified rlimit failed with errno.

Fields of SetRlimit

errno: i32kind: c_int

SettingChrootDirectory(i32, PathBuf)

chroot failed with the provided errno.

Tuple Fields of SettingChrootDirectory

0: i321: PathBuf

SettingPivotRootDirectory(i32, PathBuf)

pivot_root failed with the provided errno.

Tuple Fields of SettingPivotRootDirectory

0: i321: PathBuf

ReadFdDirEntry(Error)

There is an entry in /proc/self/fd that isn’t a valid PID.

Tuple Fields of ReadFdDirEntry

0: Error

ReadFdDir(Error)

/proc/self/fd failed to open.

Tuple Fields of ReadFdDir

0: Error

ProcFd(String)

An entry in /proc/self/fd is not an integer

Tuple Fields of ProcFd

0: String

PreservingFd(i32)

Minijail refused to preserve an FD in the inherit list of fork().

Tuple Fields of PreservingFd

0: i32

ProgramTooLarge

Program size is too large

WrongProgramAlignment

Alignment of file should be divisible by the alignment of sock_filter.

WrongProgramSize

File size should be non-zero and a multiple of sock_filter

NoCommand

The command was not found.

NoAccess

The command could not be run.

SeccompViolation(i32)

Process was killed by SIGSYS indicating a seccomp violation.

Tuple Fields of SeccompViolation

0: i32

Killed(u8)

Process was killed by a signal other than SIGSYS.

Tuple Fields of Killed

0: u8

ReturnCode(u8)

Process finished returning a non-zero code.

Tuple Fields of ReturnCode

0: u8

Wait(i32)

Failed to wait the process.

Tuple Fields of Wait

0: i32